[REF] conditional_format: split list panel and cf preview
This commit put the CF preview in a separate component.

Task: 3571796
